Suppressed a warning produced with Visual Studio
git-svn-id: http://svn.ta3d.org/yuni/trunk@2037 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/io/file/copy.cpp
@@ -60,7 +60,7 @@ namespace File
 				char* buffer = new char[size];
 				size_t numRead;
 				while ((numRead = fromFile.read(buffer, size)) != 0)
-					toFile.write(buffer, numRead);
+					toFile.write(buffer, (unsigned int) numRead);
 				delete[] buffer;
 				return Yuni::IO::errNone;
